requirements.txt
selenium==3.141.0
allure-pytest==2.9.43
pytest==6.2.4
